requirements.txt
# LLM & Agents
groq>=0.5.0
langchain-groq>=0.1.0
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.1.0

# RAG (Vector Database & Embeddings)
chromadb>=0.4.0
sentence-transformers>=2.7.0

# API & Backend
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Code Analysis
tree-sitter>=0.21.0

# Utilities
httpx>=0.27.0
pytest>=8.0.0

bandit>=1.7.0

streamlit>=1.32.0